drivers: npmx: shell: move vbusin code to the separate file
